+++ tests/test_builders/test_build_gettext.py
@@ -170,6 +170,15 @@ def test_gettext_template_msgid_order_in_sphinxpot(app):
         flags=re.DOTALL)
 
 
+@pytest.mark.sphinx('gettext', testroot='gettext-custom-output-template')
+def test_gettext_custom_output_template(app):
+    app.build(force_all=True)
+    assert (app.outdir / 'index.pot').is_file()
+
+    result = (app.outdir / 'index.pot').read_text(encoding='utf8')
+    assert "EVEN MORE DESCRIPTIVE TITLE" in result
+
+
 @pytest.mark.sphinx(
     'gettext', srcdir='root-gettext',
     confoverrides={'gettext_compact': 'documentation'})