+++ CHANGES.rst
@@ -79,6 +79,8 @@ Features added
 
 Bugs fixed
 ----------
+* #12643: Correctly support custom gettext output templates.
+  Patch by Jeremy Bowman.
 
 Testing
 -------